Progress on blog dev - adding posts from tumblr and converting them
At first I was using the date variable in the post front matter, but
jekyll build kept choking on that, but when I removed it all was well.

I've also found that  jekyll seems to be choking on my use of markup in
the .md post files, so I dumped all of my posts into a _postarchive
directory, and I'm going through one by one and removing the markup.
I'm at a pausing point for now because there are a lot of files left
and I don't want to waste time to doing that all right now.