h1 title tweak
The spirit of using the word "people" was to emphasize that I focus on the people using software and think about who I'm designing for in human terms. But I think it might just be too "cutesy" or "philosophical" and simply using "developer" is a more clear and relevant word for the work I've done.